Don't render hidden nodes
Until now we rendered all nodes of the AST and simply hid them with CSS.
However, that is not performant for large ASTs. The advantage was that
we could maintain state for nodes. I.e. when a subtree was expanded, it
was still expanded when an ancestor node was toggled.

That's now not the case anymore, but instead when you click on a node
with the shift key pressed, the whole subtree gets expanded.