Load Babel dynamically
To reduce the size of the entry app.js file, Babel is now loaded
dynamically the first time the parser is switched.

I had to remove the `npm run watch` script since it didn't seem to work
with partition-bundle. This makes developing a but more tedious, but
it's still ok. For some reason I didn't seem to be able to get a
non-minified version of the code to run.