Give more render control to parsers
Until now we assumed that an AST was composed of plain JavaScript
objects. However, that might not be always the case. Nodes may not be
simple objects or they may expose their data via methods, not plain
properties.

With this change, parsers are now in control over

  - the name of a node (`node.type` until now)
  - how to convert a node to something renderable
  - which part of the tree should be opened (by default). This helps
    especially with circular structures.

In particular, parsers must now expose a `forEachProperty` method. This
method accepts an AST node and a callback. The callback accepts an
object with the properties "value", "key", "custom" and "computed".
These tell the renderer how to represent the node and which part of it
may have been computed dynamically (e.g. via a function call).

In addition to the above changes, the node that was last toggled is now
accessible via the global `$node` variable (for inspection in the
console).